added seperate components
